Split lsp logs into "backend" and UI parts, use only the former on the remote headless server
