+++ crates/language_tools/src/lsp_button.rs
@@ -11,7 +11,10 @@ use editor::{Editor, EditorEvent};
 use gpui::{Corner, Entity, Subscription, Task, WeakEntity, actions};
 use language::{BinaryStatus, BufferId, ServerHealth};
 use lsp::{LanguageServerId, LanguageServerName, LanguageServerSelector};
-use project::{LspStore, LspStoreEvent, Worktree, project_settings::ProjectSettings};
+use project::{
+    LspStore, LspStoreEvent, Worktree, lsp_store::log_store::GlobalLogStore,
+    project_settings::ProjectSettings,
+};
 use settings::{Settings as _, SettingsStore};
 use ui::{
     Context, ContextMenu, ContextMenuEntry, ContextMenuItem, DocumentationAside, DocumentationSide,
@@ -20,7 +23,7 @@ use ui::{
 
 use workspace::{StatusItemView, Workspace};
 
-use crate::lsp_log::GlobalLogStore;
+use crate::lsp_log_view;
 
 actions!(
     lsp_tool,
@@ -30,7 +33,7 @@ actions!(
     ]
 );
 
-pub struct LspTool {
+pub struct LspButton {
     server_state: Entity<LanguageServerState>,
     popover_menu_handle: PopoverMenuHandle<ContextMenu>,
     lsp_menu: Option<Entity<ContextMenu>>,
@@ -339,14 +342,13 @@ impl LanguageServerState {
                             })
                             .detach();
                         } else if has_logs {
-                            lsp_logs.update(cx, |lsp_logs, cx| {
-                                lsp_logs.open_server_trace(
-                                    workspace.clone(),
-                                    server_selector.clone(),
-                                    window,
-                                    cx,
-                                );
-                            });
+                            lsp_log_view::open_server_trace(
+                                &lsp_logs,
+                                workspace.clone(),
+                                server_selector.clone(),
+                                window,
+                                cx,
+                            );
                         } else {
                             cx.propagate();
                         }
@@ -510,7 +512,7 @@ impl ServerData<'_> {
     }
 }
 
-impl LspTool {
+impl LspButton {
     pub fn new(
         workspace: &Workspace,
         popover_menu_handle: PopoverMenuHandle<ContextMenu>,
@@ -518,12 +520,12 @@ impl LspTool {
         cx: &mut Context<Self>,
     ) -> Self {
         let settings_subscription =
-            cx.observe_global_in::<SettingsStore>(window, move |lsp_tool, window, cx| {
+            cx.observe_global_in::<SettingsStore>(window, move |lsp_button, window, cx| {
                 if ProjectSettings::get_global(cx).global_lsp_settings.button {
-                    if lsp_tool.lsp_menu.is_none() {
-                        lsp_tool.refresh_lsp_menu(true, window, cx);
+                    if lsp_button.lsp_menu.is_none() {
+                        lsp_button.refresh_lsp_menu(true, window, cx);
                     }
-                } else if lsp_tool.lsp_menu.take().is_some() {
+                } else if lsp_button.lsp_menu.take().is_some() {
                     cx.notify();
                 }
             });
@@ -541,8 +543,8 @@ impl LspTool {
         }
 
         let lsp_store_subscription =
-            cx.subscribe_in(&lsp_store, window, |lsp_tool, _, e, window, cx| {
-                lsp_tool.on_lsp_store_event(e, window, cx)
+            cx.subscribe_in(&lsp_store, window, |lsp_button, _, e, window, cx| {
+                lsp_button.on_lsp_store_event(e, window, cx)
             });
 
         let server_state = cx.new(|_| LanguageServerState {
@@ -553,24 +555,24 @@ impl LspTool {
             language_servers,
         });
 
-        let mut lsp_tool = Self {
+        let mut lsp_button = Self {
             server_state,
             popover_menu_handle,
             lsp_menu: None,
             lsp_menu_refresh: Task::ready(()),
             _subscriptions: vec![settings_subscription, lsp_store_subscription],
         };
-        if !lsp_tool
+        if !lsp_button
             .server_state
             .read(cx)
             .language_servers
             .binary_statuses
             .is_empty()
         {
-            lsp_tool.refresh_lsp_menu(true, window, cx);
+            lsp_button.refresh_lsp_menu(true, window, cx);
         }
 
-        lsp_tool
+        lsp_button
     }
 
     fn on_lsp_store_event(
@@ -893,18 +895,18 @@ impl LspTool {
     ) {
         if create_if_empty || self.lsp_menu.is_some() {
             let state = self.server_state.clone();
-            self.lsp_menu_refresh = cx.spawn_in(window, async move |lsp_tool, cx| {
+            self.lsp_menu_refresh = cx.spawn_in(window, async move |lsp_button, cx| {
                 cx.background_executor()
                     .timer(Duration::from_millis(30))
                     .await;
-                lsp_tool
-                    .update_in(cx, |lsp_tool, window, cx| {
-                        lsp_tool.regenerate_items(cx);
+                lsp_button
+                    .update_in(cx, |lsp_button, window, cx| {
+                        lsp_button.regenerate_items(cx);
                         let menu = ContextMenu::build(window, cx, |menu, _, cx| {
                             state.update(cx, |state, cx| state.fill_menu(menu, cx))
                         });
-                        lsp_tool.lsp_menu = Some(menu.clone());
-                        lsp_tool.popover_menu_handle.refresh_menu(
+                        lsp_button.lsp_menu = Some(menu.clone());
+                        lsp_button.popover_menu_handle.refresh_menu(
                             window,
                             cx,
                             Rc::new(move |_, _| Some(menu.clone())),
@@ -917,7 +919,7 @@ impl LspTool {
     }
 }
 
-impl StatusItemView for LspTool {
+impl StatusItemView for LspButton {
     fn set_active_pane_item(
         &mut self,
         active_pane_item: Option<&dyn workspace::ItemHandle>,
@@ -940,9 +942,9 @@ impl StatusItemView for LspTool {
                     let _editor_subscription = cx.subscribe_in(
                         &editor,
                         window,
-                        |lsp_tool, _, e: &EditorEvent, window, cx| match e {
+                        |lsp_button, _, e: &EditorEvent, window, cx| match e {
                             EditorEvent::ExcerptsAdded { buffer, .. } => {
-                                let updated = lsp_tool.server_state.update(cx, |state, cx| {
+                                let updated = lsp_button.server_state.update(cx, |state, cx| {
                                     if let Some(active_editor) = state.active_editor.as_mut() {
                                         let buffer_id = buffer.read(cx).remote_id();
                                         active_editor.editor_buffers.insert(buffer_id)
@@ -951,13 +953,13 @@ impl StatusItemView for LspTool {
                                     }
                                 });
                                 if updated {
-                                    lsp_tool.refresh_lsp_menu(false, window, cx);
+                                    lsp_button.refresh_lsp_menu(false, window, cx);
                                 }
                             }
                             EditorEvent::ExcerptsRemoved {
                                 removed_buffer_ids, ..
                             } => {
-                                let removed = lsp_tool.server_state.update(cx, |state, _| {
+                                let removed = lsp_button.server_state.update(cx, |state, _| {
                                     let mut removed = false;
                                     if let Some(active_editor) = state.active_editor.as_mut() {
                                         for id in removed_buffer_ids {
@@ -971,7 +973,7 @@ impl StatusItemView for LspTool {
                                     removed
                                 });
                                 if removed {
-                                    lsp_tool.refresh_lsp_menu(false, window, cx);
+                                    lsp_button.refresh_lsp_menu(false, window, cx);
                                 }
                             }
                             _ => {}
@@ -1001,7 +1003,7 @@ impl StatusItemView for LspTool {
     }
 }
 
-impl Render for LspTool {
+impl Render for LspButton {
     fn render(&mut self, _: &mut Window, cx: &mut Context<Self>) -> impl ui::IntoElement {
         if self.server_state.read(cx).language_servers.is_empty() || self.lsp_menu.is_none() {
             return div();
@@ -1046,11 +1048,11 @@ impl Render for LspTool {
             (None, "All Servers Operational")
         };
 
-        let lsp_tool = cx.entity();
+        let lsp_button = cx.entity();
 
         div().child(
             PopoverMenu::new("lsp-tool")
-                .menu(move |_, cx| lsp_tool.read(cx).lsp_menu.clone())
+                .menu(move |_, cx| lsp_button.read(cx).lsp_menu.clone())
                 .anchor(Corner::BottomLeft)
                 .with_handle(self.popover_menu_handle.clone())
                 .trigger_with_tooltip(
